feat: CascadingMenuHelper update to winui3/release/1.5.4
